pyproject.toml
[project]
name = "lab4"
version = "0.0.0"
description = "Шаблон для создания лабораторной работы 4 по машинному обучению"
authors = [
    {name = "Your Name",email = "you@example.com"}
]

requires-python = ">=3.11,<3.13"
dependencies = [
    "gradio (>=5.28.0,<6.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.14"

